fix(engine): prevent sql errors on new user login (#1447)
If a server is ran with `Environment.WEBSITE_REGISTRATION` **disabled**,
meaning that users can login with any credentials and have an account
created, there will be some SQL errors because the response from
LoginServer.ts is missing the `account_id` property. Fix this by
returning the primary key inserted for the new account.

Co-authored-by: locked-chest <179689881+sentientflare@users.noreply.github.com>